+++ src/panels/beaconSummaryPanel/BeaconSummaryPanel.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beaconFixture } from "../../fixtures/beacons.fixture";
 import { IBeaconsGateway } from "../../gateways/IBeaconsGateway";
-import { Placeholders } from "../../useCases/mcaWritingStyleFormatter";
+import { Placeholders } from "../../utils/mcaWritingStyleFormatter";
 import { BeaconSummaryPanel } from "./BeaconSummaryPanel";
 
 describe("BeaconSummaryPanel", () => {
