+++ src/panels/beaconSummaryPanel/BeaconSummaryPanel.tsx
@@ -6,7 +6,7 @@ import { LoadingState } from "../../components/dataPanel/PanelLoadingState";
 import { DataPanelStates } from "../../components/dataPanel/States";
 import { IBeacon } from "../../entities/IBeacon";
 import { IBeaconsGateway } from "../../gateways/IBeaconsGateway";
-import { Placeholders } from "../../useCases/mcaWritingStyleFormatter";
+import { Placeholders } from "../../utils/mcaWritingStyleFormatter";
 import { BeaconSummaryEditing } from "./BeaconSummaryEditing";
 import { BeaconSummaryViewing } from "./BeaconSummaryViewing";
