+++ src/panels/beaconSummaryPanel/BeaconSummaryViewing.tsx
@@ -5,7 +5,7 @@ import {
   formatEmergencyContacts,
   formatOwners,
   formatUses,
-} from "../../useCases/mcaWritingStyleFormatter";
+} from "../../utils/mcaWritingStyleFormatter";
 
 export const BeaconSummaryViewing: FunctionComponent<{
   beacon: IBeacon;
