+++ src/utils/mcaWritingStyleFormatter.test.tsx
