+++ src/utils/mcaWritingStyleFormatter.tsx
